requirements.txt
django
python-decouple
gunicorn
whitenoise
requests
torch
numpy
ultralytics
scikit-learn
opencv-python-headless==4.8.1.78
opencv-python
paho-mqtt
django-storages 
boto3


